Try to make "TestToolService" more robust: Move "TestToolService.cs out of SkylineTool.dll, rename it "ToolServiceTestSharedClasses" and put it in "TestFunctional" folder.
