+++ pwiz_tools/Skyline/Executables/Tools/TestInteractiveTool/TestInteractiveTool.csproj
@@ -76,6 +76,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="..\..\..\TestFunctional\ToolServiceTestSharedClasses.cs">
+      <Link>ToolServiceTestSharedClasses.cs</Link>
+    </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
