+++ pwiz_tools/Skyline/SkylineTool/SkylineTool.csproj
@@ -94,7 +94,6 @@
     <Compile Include="RemoteClient.cs" />
     <Compile Include="RemoteService.cs" />
     <Compile Include="SkylineToolClient.cs" />
-    <Compile Include="TestToolService.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
